+++ lib/src/selection/material_text_selection_controls.dart
@@ -1,26 +1,25 @@
+// Copyright 2016 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'dart:math' as math;
 
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter/rendering.dart';
-
-///
-///  create by zmtzawqlp on 2019/6/10
-///
-
-/// Text selection controls that follow the Material Design specification.
-final TextSelectionControls materialExtendedTextSelectionControls =
-    MaterialExtendedTextSelectionControls();
-
 const double _kHandleSize = 22.0;
 
 // Minimal padding from all edges of the selection toolbar to all edges of the
 // viewport.
 const double _kToolbarScreenPadding = 8.0;
 const double _kToolbarHeight = 44.0;
+// Padding when positioning toolbar below selection.
+const double _kToolbarContentDistanceBelow = 16.0;
+const double _kToolbarContentDistance = 8.0;
 
 /// Manages a copy/paste text selection toolbar.
-class _TextSelectionToolbar extends StatelessWidget {
-  const _TextSelectionToolbar({
+class ExtendedMaterialTextSelectionToolbar extends StatelessWidget {
+  const ExtendedMaterialTextSelectionToolbar({
     Key key,
     this.handleCut,
     this.handleCopy,
@@ -35,25 +34,24 @@ class _TextSelectionToolbar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final List<Widget> items = <Widget>[];
     final MaterialLocalizations localizations =
         MaterialLocalizations.of(context);
-
-    if (handleCut != null)
-      items.add(FlatButton(
-          child: Text(localizations.cutButtonLabel), onPressed: handleCut));
-    if (handleCopy != null)
-      items.add(FlatButton(
-          child: Text(localizations.copyButtonLabel), onPressed: handleCopy));
-    if (handlePaste != null)
-      items.add(FlatButton(
-        child: Text(localizations.pasteButtonLabel),
-        onPressed: handlePaste,
-      ));
-    if (handleSelectAll != null)
-      items.add(FlatButton(
-          child: Text(localizations.selectAllButtonLabel),
-          onPressed: handleSelectAll));
+    final List<Widget> items = <Widget>[
+      if (handleCut != null)
+        FlatButton(
+            child: Text(localizations.cutButtonLabel), onPressed: handleCut),
+      if (handleCopy != null)
+        FlatButton(
+            child: Text(localizations.copyButtonLabel), onPressed: handleCopy),
+      if (handlePaste != null)
+        FlatButton(
+            child: Text(localizations.pasteButtonLabel),
+            onPressed: handlePaste),
+      if (handleSelectAll != null)
+        FlatButton(
+            child: Text(localizations.selectAllButtonLabel),
+            onPressed: handleSelectAll),
+    ];
 
     // If there is no option available, build an empty widget.
     if (items.isEmpty) {
@@ -72,9 +70,8 @@ class _TextSelectionToolbar extends StatelessWidget {
 
 /// Centers the toolbar around the given position, ensuring that it remains on
 /// screen.
-class MaterialExtendedTextSelectionToolbarLayout
-    extends SingleChildLayoutDelegate {
-  MaterialExtendedTextSelectionToolbarLayout(
+class ExtendedMaterialTextSelectionToolbarLayout extends SingleChildLayoutDelegate {
+  ExtendedMaterialTextSelectionToolbarLayout(
       this.screenSize, this.globalEditableRegion, this.position);
 
   /// The size of the screen at the time that the toolbar was last laid out.
@@ -114,14 +111,14 @@ class MaterialExtendedTextSelectionToolbarLayout
   }
 
   @override
-  bool shouldRelayout(MaterialExtendedTextSelectionToolbarLayout oldDelegate) {
+  bool shouldRelayout(ExtendedMaterialTextSelectionToolbarLayout oldDelegate) {
     return position != oldDelegate.position;
   }
 }
 
 /// Draws a single text selection handle which points up and to the left.
-class _TextSelectionHandlePainter extends CustomPainter {
-  _TextSelectionHandlePainter({this.color});
+class ExtendedMaterialTextSelectionHandlePainter extends CustomPainter {
+  ExtendedMaterialTextSelectionHandlePainter({this.color});
 
   final Color color;
 
@@ -134,12 +131,12 @@ class _TextSelectionHandlePainter extends CustomPainter {
   }
 
   @override
-  bool shouldRepaint(_TextSelectionHandlePainter oldPainter) {
+  bool shouldRepaint(ExtendedMaterialTextSelectionHandlePainter oldPainter) {
     return color != oldPainter.color;
   }
 }
 
-class MaterialExtendedTextSelectionControls extends TextSelectionControls {
+class ExtendedMaterialTextSelectionControls extends TextSelectionControls {
   /// Returns the size of the Material handle.
   @override
   Size getHandleSize(double textLineHeight) =>
@@ -161,33 +158,31 @@ class MaterialExtendedTextSelectionControls extends TextSelectionControls {
     // The toolbar should appear below the TextField
     // when there is not enough space above the TextField to show it.
     final TextSelectionPoint startTextSelectionPoint = endpoints[0];
-    final TextSelectionPoint endTextSelectionPoint =
-        (endpoints.length > 1) ? endpoints[1] : null;
-    final double x = (endTextSelectionPoint == null)
-        ? startTextSelectionPoint.point.dx
-        : (startTextSelectionPoint.point.dx + endTextSelectionPoint.point.dx) /
-            2.0;
-    final double availableHeight = globalEditableRegion.top -
-        MediaQuery.of(context).padding.top -
-        _kToolbarScreenPadding;
-    final double y = (availableHeight < _kToolbarHeight)
-        ? startTextSelectionPoint.point.dy +
-            globalEditableRegion.height +
+    final double toolbarHeightNeeded = MediaQuery.of(context).padding.top +
+        _kToolbarScreenPadding +
+        _kToolbarHeight +
+        _kToolbarContentDistance;
+    final double availableHeight =
+        globalEditableRegion.top + endpoints.first.point.dy - textLineHeight;
+    final bool fitsAbove = toolbarHeightNeeded <= availableHeight;
+    final double y = fitsAbove
+        ? startTextSelectionPoint.point.dy -
+            _kToolbarContentDistance -
+            textLineHeight
+        : startTextSelectionPoint.point.dy +
             _kToolbarHeight +
-            _kToolbarScreenPadding
-        : startTextSelectionPoint.point.dy - textLineHeight * 2.0;
-
-    final Offset preciseMidpoint = Offset(x, y);
+            _kToolbarContentDistanceBelow;
+    final Offset preciseMidpoint = Offset(position.dx, y);
 
     return ConstrainedBox(
       constraints: BoxConstraints.tight(globalEditableRegion.size),
       child: CustomSingleChildLayout(
-        delegate: MaterialExtendedTextSelectionToolbarLayout(
+        delegate: ExtendedMaterialTextSelectionToolbarLayout(
           MediaQuery.of(context).size,
           globalEditableRegion,
           preciseMidpoint,
         ),
-        child: _TextSelectionToolbar(
+        child: ExtendedMaterialTextSelectionToolbar(
           handleCut: canCut(delegate) ? () => handleCut(delegate) : null,
           handleCopy: canCopy(delegate) ? () => handleCopy(delegate) : null,
           handlePaste: canPaste(delegate) ? () => handlePaste(delegate) : null,
@@ -206,7 +201,7 @@ class MaterialExtendedTextSelectionControls extends TextSelectionControls {
       width: _kHandleSize,
       height: _kHandleSize,
       child: CustomPaint(
-        painter: _TextSelectionHandlePainter(
+        painter: ExtendedMaterialTextSelectionHandlePainter(
             color: Theme.of(context).textSelectionHandleColor),
       ),
     );
@@ -252,8 +247,13 @@ class MaterialExtendedTextSelectionControls extends TextSelectionControls {
     // Android allows SelectAll when selection is not collapsed, unless
     // everything has already been selected.
     final TextEditingValue value = delegate.textEditingValue;
-    return value.text.isNotEmpty &&
+    return delegate.selectAllEnabled &&
+        value.text.isNotEmpty &&
         !(value.selection.start == 0 &&
             value.selection.end == value.text.length);
   }
 }
+
+/// Text selection controls that follow the Material Design specification.
+final TextSelectionControls extendedMaterialTextSelectionControls =
+    ExtendedMaterialTextSelectionControls();